+++ src/Database/DefaultMaterializedView.stub.php
@@ -12,8 +12,7 @@ namespace Cassandra {
  */    final class DefaultMaterializedView extends \Cassandra\MaterializedView {
         public function name(): string {}
 
-        /** @return mixed */
-        public function option(string $name): mixed {}
+        public function option(string $name): Value|string|int|float|bool|null {}
 
         /** @return array<string, mixed> */
         public function options(): array {}
@@ -57,11 +56,9 @@ namespace Cassandra {
         /** @return array<string, mixed>|false */
         public function compressionParameters(): array|false {}
 
-        /** @return bool|false */
-        public function populateIOCacheOnFlush(): bool|false {}
+        public function populateIOCacheOnFlush(): bool {}
 
-        /** @return bool|false */
-        public function replicateOnWrite(): bool|false {}
+        public function replicateOnWrite(): bool {}
 
         /** @return int|false */
         public function maxIndexInterval(): int|false {}